💅 Rename Notes from Post
