+++ utils/createSlug.ts
@@ -1,4 +1,4 @@
-import type { Frontmatter } from "../types/post";
+import type { Frontmatter } from "../types/index";
 
 export function createSlug(fileName: string, frontmatter: Frontmatter) {
   return frontmatter?.slug ?? fileName.replace(".md", "");
