+++ utils/pillHelpers.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { PillProps } from "@/components/Pill";
-import { ExpandedNote } from "types/post";
+import { ExpandedNote } from "types/index";
 export type PillInput = {
   id: string;
   count?: number;
